Move FormGroupsWidget
