--- frontend/src/metabase/forms/components/FormGroupsWidget/index.ts
@@ -1,1 +1,0 @@
-export * from "./FormGroupsWidget";
